requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
PyJWT==2.9.0
sqlparse==0.5.1
asgiref==3.8.1
typing-extensions==4.12.2
psycopg2-binary==2.9.10
